requirements.txt
pandas==1.3.3
sqlalchemy==1.4.25
boto3==1.22.4
psycopg2-binary==2.9.1
PyYAML~=5.4.1
pip~=21.2.4


apache-airflow-providers-docker==2.6.0
python-dateutil==2.8.2
plyvel==1.4.0